+++ client/src/pages/SignIn.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import SignInComponent from '../components/user/SignInComponent';
+
+const SignIn = () => {
+  return (
+    <>
+        <SignInComponent/>
+    </>
+  )
+}
+
+export default SignIn;
