+++ client/src/pages/SignUp.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import SignUpComponent from '../components/user/SignUpComponent';
+
+const SignUp = () => {
+  return (
+    <>
+        <SignUpComponent/>
+    </>
+  )
+}
+
+export default SignUp;
